requirements.txt
# Core FastAPI
fastapi==0.116.1
uvicorn==0.35.0

# Base de datos
SQLAlchemy==2.0.43
asyncpg==0.29.0
psycopg2-binary==2.9.10

# Validaciones
pydantic==2.11.7
email-validator==2.2.0

# Configuraciones
pydantic-settings==2.7.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Cache y Redis
redis==5.1.1

# Logging estructurado
python-json-logger==2.0.7

# Security headers
secure==0.3.0

# System monitoring
psutil==6.0.0

# Production optimizations
uvloop==0.20.1
httptools==0.6.4

# Optional monitoring integrations
# sentry-sdk[fastapi]==1.45.1
# newrelic==9.13.0
